pyproject.toml
[project]
name = "aura-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.53.0",
    "langchain>=1.1.2",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "numpy>=1.24.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "sqlmodel>=0.0.27",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "google-search-results>=2.4.2",
    "psycopg2-binary>=2.9.11",
    "psycopg-pool>=3.2.0",
    "langgraph-checkpoint-postgres>=3.0.1",
    "supabase>=2.25.0",
    "fastapi>=0.124.0",
    "uvicorn[standard]>=0.38.0",
    "asyncpg>=0.31.0",
    "greenlet>=3.3.0",
    "boto3>=1.35.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
]
